Conflicts:
	target/m2e-wtp/web-resources/META-INF/MANIFEST.MF

+++ src/main/java/com/ryan/mapper/DetailBookMapper.java
@@ -28,7 +28,7 @@ public interface DetailBookMapper {
 	public ArrayList<BookLikeVO> bookLikeList(int booknumber);/////
 	
 	//평점 조회
-	public double bookGrade(int booknumber);
+	public int bookGrade(int booknumber);
 	
 	//태그 조회
 	public List<HashtagVO> hashtag(int booknumber);
